# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Holding Today - a private place for bereaved parents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memories:
    get:
      operationId: getMemories
      tags: [memories]
      summary: Get all memories
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memories]
      summary: Create a memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
  /memories/{id}:
    put:
      operationId: updateMemory
      tags: [memories]
      summary: Update a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /stories:
    get:
      operationId: getStories
      tags: [stories]
      summary: Get all shared stories
      responses:
        "200":
          description: List of stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Story"
    post:
      operationId: createStory
      tags: [stories]
      summary: Share a story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{id}:
    delete:
      operationId: deleteStory
      tags: [stories]
      summary: Delete a story
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Memory:
      type: object
      properties:
        isStarred:
          type: boolean
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        dateTaken:
          type: string
          nullable: true
        tags:
          type: string
          nullable: true
        isAiGenerated:
          type: boolean
        createdAt:
          type: string
      required: [id, title, isAiGenerated, createdAt]
    MemoryInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        dateTaken:
          type: string
          nullable: true
        tags:
          type: string
          nullable: true
        isAiGenerated:
          type: boolean
      required: [title]
...
    Story:
      type: object
      properties:
        id:
          type: integer
        authorName:
          type: string
        relationship:
          type: string
          nullable: true
        content:
          type: string
        title:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, authorName, content, createdAt]
    StoryInput:
      type: object
      properties:
        authorName:
          type: string
        relationship:
          type: string
          nullable: true
        content:
          type: string
        title:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
      required: [authorName, content]
...